backend/requirements.txt
click==8.1.3
colorama==0.4.5
Flask==2.1.3
Flask-Cors==3.0.10
Flask-PyMongo==2.3.0
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
PyJWT==2.4.0
pymongo==4.1.1
python-dotenv==0.20.0
six==1.16.0
Werkzeug==2.1.2
zipp==3.8.1
